Revert "#367 change selection of simple objects (#381)"
This reverts commit 92c1da66ba63f0367a4dd6d9eea456db7c46f6dc.
