+++ packages/ketcher-react/src/script/editor/tool/helper/locate.js
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2021 EPAM Systems
+ * Copyright 2020 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -71,11 +71,13 @@ function getElementsInRectangle(restruct, p0, p1) {
   })
 
   restruct.simpleObjects.forEach((item, id) => {
-    const referencePoints = item.getReferencePoints(true)
-    const referencePointInRectangle = referencePoints.find(
-      point => point.x > x0 && point.x < x1 && point.y > y0 && point.y < y1
+    if (
+      item.item.pos[0].x > x0 &&
+      item.item.pos[0].x < x1 &&
+      item.item.pos[0].y > y0 &&
+      item.item.pos[0].y < y1
     )
-    if (referencePointInRectangle) simpleObjectsList.push(id)
+      simpleObjectsList.push(id)
   })
 
   const enhancedFlagList = []
@@ -142,11 +144,7 @@ function getElementsInPolygon(restruct, rr) {
   })
 
   restruct.simpleObjects.forEach((item, id) => {
-    const referencePoints = item.getReferencePoints(true)
-    const referencePointInPolygon = referencePoints.find(point =>
-      isPointInPolygon(r, point)
-    )
-    if (referencePointInPolygon) simpleObjectsList.push(id)
+    if (isPointInPolygon(r, item.item.pos[0])) simpleObjectsList.push(id)
   })
 
   const enhancedFlagList = []
