+++ packages/ketcher-react/src/script/render/restruct/reatom.js
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2021 EPAM Systems
+ * Copyright 2020 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-import ReObject from './ReObject'
+import ReObject from './reobject'
 import element from '../../chem/element'
 import { sketchingColors as elementColor } from '../../chem/element-color'
 import draw from '../draw'
@@ -23,6 +23,7 @@ import { Bond, Box2Abs, Vec2, scale } from 'ketcher-core'
 /** @param {import('ketcher-core').Atom} atom */
 function ReAtom(/* chem.Atom*/ atom) {
   this.visel = undefined // for typing
+  this.init('atom')
 
   /** @type {import('ketcher-core').Atom} */
   this.a = atom // TODO rename a to item
@@ -34,7 +35,7 @@ function ReAtom(/* chem.Atom*/ atom) {
   this.component = -1
 }
 
-ReAtom.prototype = new ReObject('atom')
+ReAtom.prototype = new ReObject()
 ReAtom.isSelectable = function () {
   return true
 }
