+++ packages/ketcher-react/src/script/render/restruct/redatasgroupdata.js
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2021 EPAM Systems
+ * Copyright 2020 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,16 @@
  * limitations under the License.
  ***************************************************************************/
 
-import ReObject from './ReObject'
+import ReObject from './reobject'
 import { scale } from 'ketcher-core'
 
 function ReDataSGroupData(sgroup) {
+  this.init('sgroupData')
+
   this.sgroup = sgroup
 }
 
-ReDataSGroupData.prototype = new ReObject('sgroupData')
+ReDataSGroupData.prototype = new ReObject()
 ReDataSGroupData.isSelectable = function () {
   return true
 }
