+++ packages/ketcher-react/src/script/render/restruct/rerxnplus.js
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2021 EPAM Systems
+ * Copyright 2020 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,17 @@
  * limitations under the License.
  ***************************************************************************/
 
-import ReObject from './ReObject'
+import ReObject from './reobject'
 import draw from '../draw'
 import util from '../util'
 import { Box2Abs, scale } from 'ketcher-core'
 
 function ReRxnPlus(/* chem.RxnPlus*/ plus) {
+  this.init('rxnPlus')
+
   this.item = plus
 }
-ReRxnPlus.prototype = new ReObject('rxnPlus')
+ReRxnPlus.prototype = new ReObject()
 ReRxnPlus.isSelectable = function () {
   return true
 }
