+++ packages/ketcher-react/src/script/render/restruct/resgroup.js
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2021 EPAM Systems
+ * Copyright 2020 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,14 +19,16 @@ import { SGroup, Box2Abs, Vec2, Pile, scale } from 'ketcher-core'
 import draw from '../draw'
 
 import ReDataSGroupData from './redatasgroupdata'
-import ReObject from './ReObject'
+import ReObject from './reobject'
 
 const tfx = util.tfx
 
 function ReSGroup(sgroup) {
+  this.init('sgroup')
+
   this.item = sgroup
 }
-ReSGroup.prototype = new ReObject('sgroup')
+ReSGroup.prototype = new ReObject()
 ReSGroup.isSelectable = function () {
   return false
 }
